+++ source/server/options_impl.h
@@ -75,11 +75,11 @@ public:
     signal_handling_enabled_ = signal_handling_enabled;
   }
   void setCpusetThreads(bool cpuset_threads_enabled) { cpuset_threads_ = cpuset_threads_enabled; }
-  void setAllowUnkownFields(bool allow_unknown_fields) {
-    allow_unknown_fields_ = allow_unknown_fields;
+  void setAllowUnkownFields(bool allow_unknown_static_fields) {
+    allow_unknown_static_fields_ = allow_unknown_static_fields;
   }
-  void setRejectUnknownFieldsDynamic(bool reject_unknown_fields_dynamic) {
-    reject_unknown_fields_dynamic_ = reject_unknown_fields_dynamic;
+  void setRejectUnknownFieldsDynamic(bool reject_unknown_dynamic_fields) {
+    reject_unknown_dynamic_fields_ = reject_unknown_dynamic_fields;
   }
 
   // Server::Options
@@ -90,8 +90,8 @@ public:
     return config_proto_;
   }
   const std::string& configYaml() const override { return config_yaml_; }
-  bool allowUnknownFields() const override { return allow_unknown_fields_; }
-  bool rejectUnknownFieldsDynamic() const override { return reject_unknown_fields_dynamic_; }
+  bool allowUnknownStaticFields() const override { return allow_unknown_static_fields_; }
+  bool rejectUnknownDynamicFields() const override { return reject_unknown_dynamic_fields_; }
   const std::string& adminAddressPath() const override { return admin_address_path_; }
   Network::Address::IpVersion localAddressIpVersion() const override {
     return local_address_ip_version_;
@@ -130,8 +130,8 @@ private:
   std::string config_path_;
   envoy::config::bootstrap::v2::Bootstrap config_proto_;
   std::string config_yaml_;
-  bool allow_unknown_fields_{false};
-  bool reject_unknown_fields_dynamic_{false};
+  bool allow_unknown_static_fields_{false};
+  bool reject_unknown_dynamic_fields_{false};
   std::string admin_address_path_;
   Network::Address::IpVersion local_address_ip_version_;
   spdlog::level::level_enum log_level_;