+++ test/integration/integration.h
@@ -194,11 +194,12 @@ public:
   void createTestServer(const std::string& json_path, const std::vector<std::string>& port_names);
   void createGeneratedApiTestServer(const std::string& bootstrap_path,
                                     const std::vector<std::string>& port_names,
-                                    bool allow_unknown_fields, bool reject_unknown_fields_dynamic,
-                                    bool allow_lds_rejection);
+                                    bool allow_unknown_static_fields,
+                                    bool reject_unknown_dynamic_fields, bool allow_lds_rejection);
   void createApiTestServer(const ApiFilesystemConfig& api_filesystem_config,
-                           const std::vector<std::string>& port_names, bool allow_unknown_fields,
-                           bool reject_unknown_fields_dynamic, bool allow_lds_rejection);
+                           const std::vector<std::string>& port_names,
+                           bool allow_unknown_static_fields, bool reject_unknown_dynamic_fields,
+                           bool allow_lds_rejection);
 
   Event::TestTimeSystem& timeSystem() { return time_system_; }