+++ test/mocks/server/mocks.h
@@ -64,8 +64,8 @@ public:
   MOCK_CONST_METHOD0(configPath, const std::string&());
   MOCK_CONST_METHOD0(configProto, const envoy::config::bootstrap::v2::Bootstrap&());
   MOCK_CONST_METHOD0(configYaml, const std::string&());
-  MOCK_CONST_METHOD0(allowUnknownFields, bool());
-  MOCK_CONST_METHOD0(rejectUnknownFieldsDynamic, bool());
+  MOCK_CONST_METHOD0(allowUnknownStaticFields, bool());
+  MOCK_CONST_METHOD0(rejectUnknownDynamicFields, bool());
   MOCK_CONST_METHOD0(adminAddressPath, const std::string&());
   MOCK_CONST_METHOD0(localAddressIpVersion, Network::Address::IpVersion());
   MOCK_CONST_METHOD0(drainTime, std::chrono::seconds());
@@ -91,8 +91,8 @@ public:
   std::string config_path_;
   envoy::config::bootstrap::v2::Bootstrap config_proto_;
   std::string config_yaml_;
-  bool allow_unknown_fields_{};
-  bool reject_unknown_fields_dynamic_{};
+  bool allow_unknown_static_fields_{};
+  bool reject_unknown_dynamic_fields_{};
   std::string admin_address_path_;
   std::string service_cluster_name_;
   std::string service_node_name_;