+++ test/server/options_impl_test.cc
@@ -76,8 +76,8 @@ TEST_F(OptionsImplTest, All) {
       "--service-cluster cluster --service-node node --service-zone zone "
       "--file-flush-interval-msec 9000 "
       "--drain-time-s 60 --log-format [%v] --parent-shutdown-time-s 90 --log-path /foo/bar "
-      "--disable-hot-restart --cpuset-threads --allow-unknown-fields "
-      "--reject-unknown-fields-dynamic");
+      "--disable-hot-restart --cpuset-threads --allow-unknown-static-fields "
+      "--reject-unknown-dynamic-fields");
   EXPECT_EQ(Server::Mode::Validate, options->mode());
   EXPECT_EQ(2U, options->concurrency());
   EXPECT_EQ("hello", options->configPath());
@@ -97,8 +97,8 @@ TEST_F(OptionsImplTest, All) {
   EXPECT_TRUE(options->hotRestartDisabled());
   EXPECT_FALSE(options->libeventBufferEnabled());
   EXPECT_TRUE(options->cpusetThreadsEnabled());
-  EXPECT_TRUE(options->allowUnknownFields());
-  EXPECT_TRUE(options->rejectUnknownFieldsDynamic());
+  EXPECT_TRUE(options->allowUnknownStaticFields());
+  EXPECT_TRUE(options->rejectUnknownDynamicFields());
 
   options = createOptionsImpl("envoy --mode init_only");
   EXPECT_EQ(Server::Mode::InitOnly, options->mode());
@@ -159,8 +159,8 @@ TEST_F(OptionsImplTest, SetAll) {
   EXPECT_EQ(!hot_restart_disabled, options->hotRestartDisabled());
   EXPECT_EQ(!signal_handling_enabled, options->signalHandlingEnabled());
   EXPECT_EQ(!cpuset_threads_enabled, options->cpusetThreadsEnabled());
-  EXPECT_TRUE(options->allowUnknownFields());
-  EXPECT_TRUE(options->rejectUnknownFieldsDynamic());
+  EXPECT_TRUE(options->allowUnknownStaticFields());
+  EXPECT_TRUE(options->rejectUnknownDynamicFields());
 
   // Validate that CommandLineOptions is constructed correctly.
   Server::CommandLineOptionsPtr command_line_options = options->toCommandLineOptions();
@@ -211,8 +211,8 @@ TEST_F(OptionsImplTest, DefaultParams) {
   EXPECT_EQ(envoy::admin::v2alpha::CommandLineOptions::Serve, command_line_options->mode());
   EXPECT_FALSE(command_line_options->disable_hot_restart());
   EXPECT_FALSE(command_line_options->cpuset_threads());
-  EXPECT_FALSE(command_line_options->allow_unknown_fields());
-  EXPECT_FALSE(command_line_options->reject_unknown_fields_dynamic());
+  EXPECT_FALSE(command_line_options->allow_unknown_static_fields());
+  EXPECT_FALSE(command_line_options->reject_unknown_dynamic_fields());
 }
 
 // Validates that the server_info proto is in sync with the options.