+++ docs/root/intro/version_history.rst
@@ -10,7 +10,7 @@ Version history
 * config: enforcing that terminal filters (e.g. HttpConnectionManager for L4, router for L7) be the last in their respective filter chains.
 * buffer filter: the buffer filter populates content-length header if not present, behavior can be disabled using the runtime feature `envoy.reloadable_features.buffer_filter_populate_content_length`.
 * config: added access log :ref:`extension filter<envoy_api_field_config.filter.accesslog.v2.AccessLogFilter.extension_filter>`.
-* config: added support for :option:`--reject-unknown-fields-dynamic`, providing independent control
+* config: added support for :option:`--reject-unknown-dynamic-fields`, providing independent control
   over whether unknown fields are rejected in static and dynamic configuration. By default, unknown
   fields in static configuration are rejected and are allowed in dynamic configuration.
 * config: async data access for local and remote data source.